Floating-point math for net worth calculation
Use floating point math for the calculation of a user's net worth, so that the exchange rate is properly applied.

+++ client.py
@@ -1699,7 +1699,7 @@ async def on_message(message):
 			market_data = EwMarket(id_server=server.id)
 
 			user_slimecredit = EwUser(member=message.author).slimecredit
-			net_worth = user_slimecredit*(market_data.rate_exchange/1000)
+			net_worth = int(user_slimecredit * (market_data.rate_exchange / 1000.0))
 			response = "You have {:,} SlimeCoin, currently valued at {:,} slime.".format(user_slimecredit, net_worth)
 
 			# Send the response to the player.
